[testing] finish and optimize unit testing

+++ app/src/main/java/com/didik/moflix/domain/usecase/SeriesUseCase.kt
@@ -38,11 +38,13 @@ class SeriesUseCase @Inject constructor(
         }
     }
 
-    suspend fun insertSeries(series: MovieModel) {
+    suspend fun insertFavoriteSeries(series: MovieModel?) {
+        if (series == null) return
         repository.insertFavoriteSeries(series)
     }
 
-    suspend fun deleteSeries(series: MovieModel) {
+    suspend fun deleteFavoriteSeries(series: MovieModel?) {
+        if (series == null) return
         repository.deleteFavoriteSeries(series)
     }
